requirements.txt
-i https://pypi.tuna.tsinghua.edu.cn/simple
grpcio==1.34.0
grpcio-tools==1.34.0
protobuf==3.19.3
msgpack==1.0.2
sentry-sdk==1.5.6
numpy>=1.21.5
pandas==1.2.4
matplotlib==3.5.1
matplotlib-inline==0.1.3
torch==1.11.0
torchvision==0.12.0
